EventClass moved to VEGAS-side
